towards arquillian tests
* wegas-app unit tests with arquillian
* reduce facades lookup usage
* get rid of useless glassfish directories

+++ wegas-app/src/test/java/com/wegas/integration/Wegas.java
@@ -117,7 +117,7 @@ public class Wegas {
 
         String root = ".";
 
-        File domainConfig = new File(root + "/src/test/glassfish/domains/domain1/config/domain.xml");
+        File domainConfig = new File(root + "/src/test/resources/domain_integration.xml");
         File tmpDomainConfig = File.createTempFile("domain", "xml");
         FileUtils.copyFile(domainConfig, tmpDomainConfig);
 
@@ -156,13 +156,15 @@ public class Wegas {
     }
 
     public static void shutdown(WegasRuntime runtime) throws GlassFishException {
-        GlassFish payara = runtime.getPayara();
-        if (payara != null) {
-            String appName = runtime.getAppName();
-            payara.getDeployer().undeploy(appName);
-            payara.stop();
+        if (runtime != null) {
+            GlassFish payara = runtime.getPayara();
+            if (payara != null) {
+                String appName = runtime.getAppName();
+                payara.getDeployer().undeploy(appName);
+                payara.stop();
+            }
+            runtime.getDomainConfig().delete();
         }
-        runtime.getDomainConfig().delete();
     }
 
 }
